Divide explosion prefabs into trigger<->consequence paradigm
