+++ Assets/Prefabs/KillZones/ExplodingGasTankMedium.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 23282731045362244}
   - component: {fileID: 64078436170890602}
   - component: {fileID: 7081591432571808861}
+  - component: {fileID: 8986445452016976163}
   m_Layer: 0
   m_Name: ExplodingGasTankMedium
   m_TagString: exploding_tank
@@ -116,3 +117,19 @@ MonoBehaviour:
   ExplosionPrefab: {fileID: 1290715649332758, guid: 587eaecaa621be74486f351c78b4f0a8, type: 3}
   ExampleCamera: {fileID: 0}
   dontRespawn: 0
+--- !u!114 &8986445452016976163
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1058322186698390}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c2c2c0e7cbc659c409600894c9d6a686, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  CustomTag: Player
+  RequireItem: 
+  onStay: 0
