+++ Assets/Prefabs/KillZones/ExplodingGasTankTall.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: 64767440466019546}
   - component: {fileID: 4683986006286791251}
   - component: {fileID: 1286524508967444037}
+  - component: {fileID: 2356844285896902660}
   m_Layer: 0
   m_Name: ExplodingGasTankTall
   m_TagString: exploding_tank
@@ -158,3 +159,19 @@ MonoBehaviour:
   Ship_Alarm: {fileID: 0}
   Turret_Firing: {fileID: 0}
   Turret_Firing_Tail: {fileID: 0}
+--- !u!114 &2356844285896902660
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1662120072097882}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c2c2c0e7cbc659c409600894c9d6a686, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  CustomTag: Player
+  RequireItem: 
+  onStay: 0
