+++ Assets/Prefabs/KillZones/ExplodingHealthPack1.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: 5033351025091367117}
   - component: {fileID: 530119411053940986}
   - component: {fileID: 6065434533076636188}
+  - component: {fileID: 3478074548994490098}
   m_Layer: 0
   m_Name: ExplodingHealthPack1
   m_TagString: Untagged
@@ -154,6 +155,22 @@ MonoBehaviour:
   Ship_Alarm: {fileID: 0}
   Turret_Firing: {fileID: 0}
   Turret_Firing_Tail: {fileID: 0}
+--- !u!114 &3478074548994490098
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1823087282522076}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c2c2c0e7cbc659c409600894c9d6a686, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  CustomTag: Player
+  RequireItem: 
+  onStay: 0
 --- !u!1 &6100892989168044031
 GameObject:
   m_ObjectHideFlags: 0
